refactor: remove `EventDetailSectionButton` component and move to `eventDetailPopup` component
